Moved 'scc' folder to 'algorithms' folder
