+++ src-tauri/src/algorithms/mod.rs
@@ -32,6 +32,9 @@ mod asymptotic_behaviour_counter;
 /// only remembers whether the behaviour was seen.
 mod asymptotic_behaviour_classifier;
 
+
+pub mod scc;
+
 // Re-export stuff from private modules to public scope as part of `algorithms` module:
 
 pub use asymptotic_behaviour_classifier::AsymptoticBehaviourClassifier;
