+++ src-tauri/src/algorithms/scc/_impl_class.rs
