+++ src-tauri/src/algorithms/scc/algo_interleaved_transition_guided_reduction/_impl_fwd_bwd_process.rs
@@ -1,7 +1,7 @@
-use crate::scc::algo_interleaved_transition_guided_reduction::{
+use crate::algorithms::scc::algo_interleaved_transition_guided_reduction::{
     BwdProcess, FwdProcess, Process, Scheduler,
 };
-use crate::scc::algo_saturated_reachability::reachability_step;
+use crate::algorithms::scc::algo_saturated_reachability::reachability_step;
 use biodivine_lib_param_bn::biodivine_std::traits::Set;
 use biodivine_lib_param_bn::symbolic_async_graph::{GraphColoredVertices, SymbolicAsyncGraph};
