+++ src-tauri/src/algorithms/scc/algo_interleaved_transition_guided_reduction/_impl_scheduler.rs
@@ -1,4 +1,4 @@
-use crate::scc::algo_interleaved_transition_guided_reduction::{Process, Scheduler};
+use crate::algorithms::scc::algo_interleaved_transition_guided_reduction::{Process, Scheduler};
 use crate::GraphTaskContext;
 use biodivine_lib_param_bn::biodivine_std::traits::Set;
 use biodivine_lib_param_bn::symbolic_async_graph::{GraphColoredVertices, SymbolicAsyncGraph};
