+++ src-tauri/src/algorithms/scc/algo_interleaved_transition_guided_reduction/mod.rs
